+++ Lunar/Fetchers/JWT.swift
@@ -11,8 +11,20 @@ import SwiftUI
 
 class JWT {
   @Default(.appBundleID) var appBundleID
+  @Default(.activeAccount) var activeAccount
 
-  func getJWTFromKeychain(actorID: String) -> String? {
+  func getJWTForActiveAccount() -> String? {
+    if let keychainObject = KeychainHelper.standard.read(
+      service: appBundleID, account: activeAccount.actorID
+    ) {
+      let jwt = String(data: keychainObject, encoding: .utf8) ?? ""
+      return jwt.replacingOccurrences(of: "\"", with: "")
+    } else {
+      return nil
+    }
+  }
+  
+  func getJWT(actorID: String) -> String? {
     if let keychainObject = KeychainHelper.standard.read(
       service: appBundleID, account: actorID
     ) {
