+++ Lunar/Fetchers/PostsFetcher.swift
@@ -44,7 +44,7 @@ class PostsFetcher: ObservableObject {
       limitParameter: 50,
       communityID: communityID,
       personID: personID,
-      jwt: getJWTFromKeychain(),
+      jwt: JWT().getJWTForActiveAccount(),
       instance: instance
     ).buildURL()
   }
@@ -105,7 +105,7 @@ class PostsFetcher: ObservableObject {
     let cacher = ResponseCacher(behavior: .cache)
 
     var headers: HTTPHeaders = []
-    if let jwt = getJWTFromKeychain() {
+    if let jwt = JWT().getJWTForActiveAccount() {
       headers = [.authorization(bearerToken: jwt)]
     }
 
@@ -222,17 +222,6 @@ class PostsFetcher: ObservableObject {
         print("PostsFetcher ERROR: \(error): \(error.errorDescription ?? "")")
         self.isLoading = false
       }
-    }
-  }
-
-  func getJWTFromKeychain() -> String? {
-    if let keychainObject = KeychainHelper.standard.read(
-      service: appBundleID, account: activeAccount.actorID
-    ) {
-      let jwt = String(data: keychainObject, encoding: .utf8) ?? ""
-      return jwt.replacingOccurrences(of: "\"", with: "")
-    } else {
-      return nil
     }
   }
 }
