+++ Lunar/Settings Views/SettingsQuicklinksView.swift
@@ -35,16 +35,6 @@ struct SettingsQuicklinksView: View {
   var iconList: [String] = CircleFillIcons().iconsList()
 
   var body: some View {
-    /// **Future implementation**
-    //    DroppableList("Users 1", users: $users1) { dropped, index in
-    //      users1.insert(dropped, at: index)
-    //      users2.removeAll { $0 == dropped }
-    //    }
-    //    DroppableList("Users 2", users: $users2)  { dropped, index in
-    //      users2.insert(dropped, at: index)
-    //      users1.removeAll { $0 == dropped }
-    //    }
-
     List {
       Section {
         Toggle(isOn: $enableQuicklinks) {
@@ -178,19 +168,6 @@ struct SettingsQuicklinksView: View {
         SortPicker(sortType: $quicklinkSort)
           .labelStyle(TitleOnlyLabelStyle())
 
-//        Picker("Post Sort", selection: $quicklinkSort) {
-//          Text("Active").tag("Active")
-//          Text("Hot").tag("Hot")
-//          Text("New").tag("New")
-//          Text("Top Day").tag("TopDay")
-//          Text("Top Week").tag("TopWeek")
-//          Text("Top Month").tag("TopMonth")
-//          Text("Top Year").tag("TopYear")
-//          Text("Top All").tag("TopAll")
-//          Text("Most Comments").tag("MostComments")
-//          Text("New Comments").tag("NewComments")
-//        }
-//        .pickerStyle(.menu)
       } header: {
         Text("Type")
       }
